requirements.txt
Django==5.2.3
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
django-cors-headers==4.7.0
psycopg2-binary==2.9.10
python-dotenv==1.0.0
django-environ==0.12.0
Pillow==10.4.0
celery==5.3.4
redis==5.0.1
django-filter==24.1
drf-yasg==1.21.7 